+++ workspaces/core/src/accordion/Accordion.tsx
@@ -1,15 +1,19 @@
-import React, { createContext, useContext, useRef } from 'react';
+import React, { useRef } from 'react';
 import * as RadixAccordion from '@radix-ui/react-accordion';
 import mergeClassnames from '../mergeClassnames/mergeClassnames';
-import AccordionItemState from './private/types/AccordionItemState';
-import AccordionProps from './private/types/AccordionProps';
-import AccordionRootProps from './private/types/AccordionRootProps';
-import AccordionSizes from './private/types/AccordionSizes';
-import ButtonProps from './private/types/ButtonProps';
-import ItemProps from './private/types/ItemProps';
+import AccordionContext from './private/utils/AccordionContext';
+import AccordionItemContext from './private/utils/AccordionItemContext';
 import setFont from './private/utils/setFont';
 import setMargin from './private/utils/setMargin';
 import setPadding from './private/utils/setPadding';
+import useAccordionContext from './private/utils/useAccordionContext';
+import useAccordionItemContext from './private/utils/useAccordionItemContext';
+import type AccordionProps from './private/types/AccordionProps';
+import type AccordionRootProps from './private/types/AccordionRootProps';
+import type ButtonProps from './private/types/ButtonProps';
+import type ContentProps from './private/types/ContentProps';
+import type HeaderProps from './private/types/HeaderProps';
+import type ItemProps from './private/types/ItemProps';
 
 const AccordionSingle = ({
   children,
@@ -40,20 +44,6 @@ const AccordionMultiple = ({
   </RadixAccordion.Root>
 );
 
-const AccordionContext = createContext<{ itemSize?: AccordionSizes }>({});
-AccordionContext.displayName = 'AccordionContext';
-
-const useAccordionContext = (component: string) => {
-  const context = useContext(AccordionContext);
-  if (context === null) {
-    const err = new Error(
-      `<${component}> is missing a parent <Accordion /> component.`
-    );
-    throw err;
-  }
-  return context;
-};
-
 const AccordionRoot = ({
   children,
   itemSize = 'md',
@@ -73,20 +63,6 @@ const AccordionRoot = ({
       </AccordionContext.Provider>
     </RadixAccordionRoot>
   );
-};
-
-const AccordionItemContext = createContext<AccordionItemState>({});
-AccordionItemContext.displayName = 'AccordionItemContext';
-
-const useAccordionItemContext = (component: string) => {
-  const context = useContext(AccordionItemContext);
-  if (context === null) {
-    const err = new Error(
-      `<${component}> is missing a parent <Accordion.Item /> component.`
-    );
-    throw err;
-  }
-  return context;
 };
 
 const Item: React.FC<ItemProps> = ({
@@ -113,26 +89,16 @@ const Item: React.FC<ItemProps> = ({
   );
 };
 
-type HeaderProps = {
-  className?: string;
-  disabled?: boolean;
-  children?: React.ReactNode;
-};
-
-const Header: React.FC<HeaderProps> = ({ children, className }) => {
-  const { contentElement } = useAccordionItemContext('Accordion.Header');
-  return (
-    <RadixAccordion.Header
-      className={mergeClassnames(
-        'w-full rounded-moon-s-sm bg-gohan',
-        contentElement?.current && 'moon-open:rounded-b-none',
-        className
-      )}
-    >
-      {children}
-    </RadixAccordion.Header>
-  );
-};
+const Header: React.FC<HeaderProps> = ({ children, className }) => (
+  <RadixAccordion.Header
+    className={mergeClassnames(
+      'w-full rounded-moon-s-sm bg-gohan moon-open:rounded-b-none',
+      className
+    )}
+  >
+    {children}
+  </RadixAccordion.Header>
+);
 
 const Button: React.FC<ButtonProps> = ({ children, className }) => {
   const { size, disabled } = useAccordionItemContext('Accordion.Button');
@@ -142,8 +108,8 @@ const Button: React.FC<ButtonProps> = ({ children, className }) => {
       className={mergeClassnames(
         setPadding(true, size),
         setFont(size),
-        'w-full justify-between flex items-center relative gap-1 cursor-pointer flex-1 font-medium text-bulma text-start',
-        'data-[disabled]:cursor-not-allowed data-[disabled]:opacity-30 ',
+        'w-full justify-between flex items-center relative gap-1 cursor-pointer flex-1 font-medium',
+        'text-bulma text-start data-[disabled]:cursor-not-allowed data-[disabled]:opacity-30 ',
         className
       )}
     >
@@ -152,17 +118,13 @@ const Button: React.FC<ButtonProps> = ({ children, className }) => {
   );
 };
 
-type ContentProps = {
-  className?: string;
-  children?: React.ReactNode;
-};
-
 const Content: React.FC<ContentProps> = ({ children, className }) => {
   const { contentElement } = useAccordionItemContext('Accordion.Content');
   return (
     <RadixAccordion.Content
       className={mergeClassnames(
-        'bg-gohan rounded-moon-s-sm rounded-t-none p-2 py-2 ps-3 pe-2 pt-0 text-moon-14 overflow-hidden w-full text-bulma',
+        'moon-open:flex bg-gohan rounded-moon-s-sm rounded-t-none p-2 py-2 ps-3 pe-2 pt-0',
+        'text-moon-14 w-full text-bulma',
         className
       )}
       ref={contentElement}
@@ -178,7 +140,7 @@ const ContentOutside: React.FC<ContentProps> = ({ children, className }) => {
     <RadixAccordion.Content
       className={mergeClassnames(
         setMargin(size),
-        'overflow-hidden w-full text-bulma text-moon-14',
+        'moon-open:flex w-full text-bulma text-moon-14',
         className
       )}
     >