+++ workspaces/core/src/accordion/private/types/HeaderProps.ts
@@ -1,0 +1,7 @@
+type HeaderProps = {
+  className?: string;
+  disabled?: boolean;
+  children?: React.ReactNode;
+};
+
+export default HeaderProps;