+++ workspaces/core/src/accordion/private/utils/AccordionContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type AccordionSizes from '../types/AccordionSizes';
+
+const AccordionContext = createContext<{ itemSize?: AccordionSizes }>({});
+AccordionContext.displayName = 'AccordionContext';
+
+export default AccordionContext;