+++ workspaces/core/src/private/icons/GenericSearch.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+const GenericSearch = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    width="1em"
+    height="1em"
+    viewBox="0 0 32 32"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M25 25l-4.413-4.405m0 0A7.962 7.962 0 109.336 9.327a7.962 7.962 0 0011.251 11.268z"
+      stroke="currentColor"
+      strokeLinecap="round"
+    />
+  </svg>
+);
+export default GenericSearch;